Operator Respawn System
Operators can reploy a short while after being taken out.

Weapons work funny after being deployed though

+++ Assets/SerializationScripts/CharStatModDictionary.cs
@@ -7,7 +7,7 @@ public class CharStatModifier
     public int health_regen = 0;
     public float regen_rate = 1f;
     public int shield_boost = 0;
-    public int speed_scale = 1;
+    public float speed_scale = 1;
     public bool speed_scale_decay = false;
     public void ApplyStats(Character target, float stat_duration, AbilityEffectComponent effect_controller = null)
     {
